robots.txt
create robots controller that checks (using an env) if it's on production or not. By default it allows everything, and you have to specify to disallow staging and acceptance. the reason for this is, when you make a mistake it let's severe than when you disallow on default and forget to set production env